+++ src/shapes/world.jl
@@ -1,16 +1,17 @@
-
 #---------------------------------------------------------
 # World type
 #---------------------------------------------------------
 """
     struct World
 
-A struct representing a collection of shapes in a 3D world.
+A struct representing a collection of shapes (and lights) in a 3D world.
 # Fields
-- `shapes::Vector{Shapes}`: the vector containing the shapes in the world.
+- `shapes::Vector{AbstractShape}`: the vector containing the shapes in the world.
+- `lights::Vector{AbstractLight}`: the vector containing the light sources in the world.
 # Constructor
 - `World()`: creates a new `World` with an empty vector of shapes.
-- `World(S::Vector{Shapes})`: creates a new `World` with the specified vector of shapes.
+- `World(S::Vector{AbstractShape})`: creates a new `World` with the specified vector of shapes. Lights are initialized to an empty vector.
+- `World(S::Vector{AbstractShape}, L::Vector{AbstractLight})`: creates a new `World` with the specified vector of shapes and light sources.
 # See also
 - [`AbstractShape`](@ref): the abstract type for all shapes.
 - [`Sphere`](@ref): a concrete implementation of `AbstractShape` representing a sphere.
@@ -18,13 +19,14 @@ A struct representing a collection of shapes in a 3D world.
 """
 struct World
     shapes::Vector{AbstractShape}
+    lights::Vector{AbstractLight}
 
-    function World()
-        new(Vector{AbstractShape}(nothing))
+    function World(shapes::Vector{AbstractShape})
+        World(shapes, Vector{AbstractLight}())
     end
-
-    function World(S::Vector{AbstractShape})
-        new(S)
+    function World(shapes::Vector{AbstractShape}, lights::Vector{AbstractLight})
+        @debug "Creating World with shapes: $(length(shapes)) and lights: $(length(lights))" shapes=shapes lights=lights
+        new(shapes, lights)
     end
 end
 
@@ -53,4 +55,28 @@ function ray_intersection(W::World, ray::Ray)
     end
 
     return closest
+end
+
+"""
+    is_point_visible(W::World, pos::Point, observer::Point)
+Checks if a point is visible from an observer's position in the world.
+See also [`is_light_visible`](@ref).
+# Arguments
+- `W::World`: the world containing the shapes
+- `pos::Point`: the position of the point to check visibility for
+- `observer::Point`: the position of the observer
+# Returns
+If the point is visible from the observer's position, returns `true`. Otherwise, returns `false`.
+"""
+function is_point_visible(W::World, pos::Point, observer::Point)
+    direction = pos - observer
+    dir_norm = norm(direction)
+    ray = Ray(origin=observer, dir=direction, tmin=1e-2 / dir_norm, tmax=1.0)
+
+    for shape in W.shapes
+        if quick_ray_intersection(shape, ray)
+            return false
+        end
+    end
+    return true
 end